support showing a loading spinner and error tooltip on ActionItem during command execution (#1448)
* support showing a loading spinner on ActionItem during command execution

If the new ActionItem prop showLoadingSpinnerDuringExecution is true, then a loading spinner is shown after the button is pressed if it executes a command. This will be used by hover buttons.

* ActionItem: support showing command execution's error in a tooltip

This adds support for showing the error inline on the button that executed the command. This is useful for errors that are scoped to a specific button instead of being global.

* add tests