Fix VS Code extension issue by sending a different GQL query for VSCE (#40621)
A [recent change](https://github.com/sourcegraph/sourcegraph/pull/40263)
changed one of the GraphQL queries that the VS Code extension is using
to add a new field that was only recently added to the GraphQL backend
(c.f.
[here](https://sourcegraph.slack.com/archives/C01LZKLRF0C/p1660889509990719)).

This unfortunately is not available on "older" server versions including
`3.42.1`. This means that every person that updates their VS Code
extension release now and is on an on-prem server that lags a bit behind
will see the search broken.

We do not yet have a system to get the server version and make decisions
based on it so I decide to fix it by reverting to the previous query in
the VS Code environment (the backend is written in a way that it can
handle older clients, just not the other way around).